// backend/src/enums/httpStatus.enums.ts
export enum HttpStatus {
  // Success Codes
  OK = 200,
  CREATED = 201,
  ACCEPTED = 202,




  // Client Errors 
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  CONFLICT = 409,

  // Server Errors 
  INTERNAL_SERVER_ERROR = 500,
  NOT_IMPLEMENTED = 501,
  BAD_GATEWAY = 502,
  SERVICE_UNAVAILABLE = 503
}